MonkaS

+++ loadbalancer.py
@@ -1,53 +1,67 @@
 from typing import List, Union
 from digitalocean import Droplet
 from docker import Container
-import os, logging, subprocess
+import os, logging
 
 class Loadbalancer:
-    def __init__(self, host : str) -> None:
-        self.host = host
+    def __init__(self, loadbalancers : List[str]) -> None:
+        self.loadbalancers = loadbalancers
 
-    def indent_config(self, config : List[str]):
-        parsed_config = ''
+    def indent(self, config : List[str]):
+        output = ''
         indentation = 0
 
         for x in config:
             if x == '}': indentation -= 1
-            parsed_config += ('\t' * indentation) + x + '\n'
+            output += ('\t' * indentation) + x + '\n'
             if x == '{': indentation += 1
 
-        return parsed_config
+        return output
 
-    def reload_nginx(self, node : Union[Droplet, Container]):
-        host = self.host if type(node) == Droplet else node.droplet.private_ip
-        os.system(f'/bin/ssh -i ~/.ssh/id_rsa -o "UserKnownHostsFile /dev/null" -o "StrictHostKeyChecking no" cloudman@{host} "sudo systemctl reload nginx"')
+    # Clean later.
+    def reload(self, node : Union[Droplet, Container]):
+        if type(node) == Container:
+            os.system(f'/bin/ssh -i ~/.ssh/id_rsa -o "UserKnownHostsFile /dev/null" -o "StrictHostKeyChecking no" cloudman@{node.droplet.private_ip} "sudo systemctl reload nginx"')
+        elif type(node) == Droplet:
+            for host in self.loadbalancers:
+                os.system(f'/bin/ssh -i ~/.ssh/id_rsa -o "UserKnownHostsFile /dev/null" -o "StrictHostKeyChecking no" cloudman@{host} "sudo systemctl reload nginx"')
+        
+        # Loadbalance the loadbalancer later.
 
-    def add_domain(self, domain : str, node : Union[Droplet, Container]):
-        path = f'nginx/main-loadbalancer/{domain}' if type(node) == Droplet else f'nginx/droplet-loadbalancer/{domain}'
+    def register_droplet(self, droplet : Droplet):
+        path = f'nginx/droplet-loadbalancer/{droplet.id}'
+        if not os.path.exists(path):
+            os.mkdir(path)
 
-        mode = 'w+' if not os.path.exists(path) else 'r+'
-        with open(path, mode) as f:
-            config = f.read().split('\n')
-            if len(config) > 1:
-                config.insert(2, f'\tserver {node.private_ip};')
-            else:
-                config = [f'upstream {domain}', '{', f'server {node.private_ip};', '}', 'server', '{', f'server_name {domain} www.{domain};', 'location /', '{', f'proxy_pass http://{domain};', '}', '}']
+    def add(self, domain : str, node : Union[Droplet, Container]):
+        if type(node) == Droplet or type(node) == Container:
+            path = f'nginx/main-loadbalancer/{domain}' if type(node) == Droplet else f'nginx/droplet-loadbalancer/{node.droplet.id}/{domain}'
+            mode = 'w+' if not os.path.exists(path) else 'r+'
 
-            f.seek(0)
-            f.write(self.indent_config(config))
-            f.truncate()
+            with open(path, mode) as f:
+                if (config := f.read().split('\n')) and mode == 'r+':
+                    config.insert(2, f'server {node.private_ip};')
+                    config = self.indent(config)
+                else:
+                    with open('nginx/example.conf', 'r') as tmp:
+                        config = tmp.read().replace('example.com', domain).replace('127.0.0.1', node.private_ip)
 
-        self.reload_nginx(node)
+                f.seek(0)
+                f.write(config)
+                f.truncate()
 
-    def remove_domain(self, domain : str, node : Union[Droplet, Container]):
-        path = f'nginx/main-loadbalancer/{domain}' if type(node) == Droplet else f'nginx/droplet-loadbalancer/{domain}'
+        self.reload(node)
 
-        with open(path, 'r+') as f:
-            config = f.read().split('\n')
-            config = [x for x in config if x != f'server {node.private_ip};']
+    def remove(self, domain : str, node : Union[Droplet, Container]):
+        if type(node) == Droplet or type(node) == Container:
+            path = f'nginx/main-loadbalancer/{domain}' if type(node) == Droplet else f'nginx/droplet-loadbalancer/{domain}'
 
-            f.seek(0)
-            f.write(self.indent_config(config))
-            f.truncate()
+            with open(path, 'r+') as f:
+                config = f.read().split('\n')
+                config = [x for x in config if x != f'server {node.private_ip};']
 
-        self.reload_nginx(node)
+                f.seek(0)
+                f.write(self.indent_config(config))
+                f.truncate()
+
+        self.reload(node)
